Move ChoicePersonalStatementComponent to provider_interface directory
  Remove editable features, PersonalStatement is only editable in the
  support interface
